This commit is an update to the database. It is set up with a few of the values and things, just a basic setup. In depth detail will be worked on.
