+++ main/admin.py
@@ -1,3 +1,6 @@
 from django.contrib import admin
+from .models import Users, WishlistItem
 
 # Register your models here.
+admin.site.register(Users);
+admin.site.register(WishlistItem);
