+++ main/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 # Create your models here.
 class Users(models.Model):
-    name = models.CharField(max_length=200)
+    name = models.CharField(max_length=20)
 
     def __str__(self):
         return self.name
