+++ main/templates/home.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Wishlist</title>
+    <link rel="stylesheet" href="stylesheet.css">
+</head>
+<body>
+<h1>Wishlist</h1>
+
+<div class="login-button-container">
+<button class="login-button">Login</button>
+</div>
+
+<div class="item-holder">
+<div class="item-div">
+    Placeholder
+</div>
+<div class="item-div">
+    Placeholder
+</div>
+</div>
+</body>
+</html>
