+++ main/views.py
@@ -1,5 +1,13 @@
 from django.shortcuts import render
+from django.http import HttpResponse
 
 # Create your views here.
-def index(response, id):
-    return HttpResponse("<h1>%s</h1>" %  id);
+
+def home(request):
+    return HttpResponse('Home')
+
+def products(request):
+    return HttpResponse('products')
+
+def customer(request):
+    return HttpResponse('customer')
